MySQL-29

+++ MySQL/mysql/innobase/data/data0data.c
@@ -193,7 +193,8 @@ dfield_check_typed_no_assert(
 
 		fprintf(stderr,
 "InnoDB: Error: data field type %lu, len %lu\n",
-			dfield_get_type(field)->mtype, dfield_get_len(field));
+			(ulong) dfield_get_type(field)->mtype,
+			(ulong) dfield_get_len(field));
 		return(FALSE);
 	}
 
@@ -215,7 +216,7 @@ dtuple_check_typed_no_assert(
 	if (dtuple_get_n_fields(tuple) > REC_MAX_N_FIELDS) {
 		fprintf(stderr,
 "InnoDB: Error: index entry has %lu fields\n",
-			dtuple_get_n_fields(tuple));
+			(ulong) dtuple_get_n_fields(tuple));
 	dump:
 		fputs("InnoDB: Tuple contents: ", stderr);
 		dtuple_print(stderr, tuple);
@@ -250,7 +251,8 @@ dfield_check_typed(
 
 		fprintf(stderr,
 "InnoDB: Error: data field type %lu, len %lu\n",
-			dfield_get_type(field)->mtype, dfield_get_len(field));
+			(ulong) dfield_get_type(field)->mtype,
+			(ulong) dfield_get_len(field));
 
 		ut_error;
 	}
@@ -444,10 +446,10 @@ dtuple_print(
 
 	n_fields = dtuple_get_n_fields(tuple);
 
-	fprintf(f, "DATA TUPLE: %lu fields;\n", n_fields);
+	fprintf(f, "DATA TUPLE: %lu fields;\n", (ulong) n_fields);
 
 	for (i = 0; i < n_fields; i++) {
-		fprintf(f, " %lu:", i);
+		fprintf(f, " %lu:", (ulong) i);
 
 		field = dtuple_get_nth_field(tuple, i);
 		
@@ -502,7 +504,7 @@ dtuple_convert_big_rec(
 
 	if (size > 1000000000) {
 		fprintf(stderr,
-"InnoDB: Warning: tuple size very big: %lu\n", size);
+"InnoDB: Warning: tuple size very big: %lu\n", (ulong) size);
 		fputs("InnoDB: Tuple contents: ", stderr);
 		dtuple_print(stderr, entry);
 		putc('\n', stderr);
